requirements-dev.txt
# requirements-dev.txt

pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
black==22.6.0
flake8==4.0.1
isort==5.10.1